Add site list endpoint
